fix: course/requirements.txt, blender tutorial 7
the Requirements.txt was never checked in properly when this submodule
was created. In advertantly the .txt files were pushed into git
attributes. This should only have been done for the glove files in the
course repo instead of at the top, so when this was copied, only the sha
was kept and the text was lost. Since it is unlikely i will go back to
coursera after all these years, I found a repo and just copied a list of
jupyter and other things as a placeholder.

This also includes a bump to the middle of Blender tutorial 5